+++ spec/core/cli_wrappers/terraform_cli_spec.rb
@@ -65,7 +65,7 @@ module Covalence
     end
 
     it "#terraform_init" do
-      expected_args = [ENV, "terraform init -get=false -input=false", anything]
+      expected_args = [ENV, "terraform init -get=false -input=false -plugin-dir=#{Covalence::TERRAFORM_PLUGIN_CACHE}", anything]
       expect(PopenWrapper).to receive(:spawn_subprocess).with(*expected_args).and_return(0)
       expect(described_class.terraform_init).to be true
     end